requirements.txt
Babel==2.8.0
Django==2.2
django-phonenumber-field==4.0.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
phonenumbers==8.12.0
Pillow==7.0.0
pytz==2019.3
sqlparse==0.3.1
